+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/edc/logic/service/EdcContractMappingService.java
@@ -20,7 +20,9 @@
 
 package org.eclipse.tractusx.puris.backend.common.edc.logic.service;
 
-import lombok.extern.slf4j.Slf4j;
+import java.lang.reflect.Constructor;
+
+import org.eclipse.tractusx.puris.backend.common.edc.domain.model.AssetType;
 import org.eclipse.tractusx.puris.backend.common.edc.domain.model.ContractMapping;
 import org.eclipse.tractusx.puris.backend.common.edc.domain.model.DtrContractMapping;
 import org.eclipse.tractusx.puris.backend.common.edc.domain.model.AssetType;
@@ -31,6 +33,7 @@ import org.eclipse.tractusx.puris.backend.common.edc.domain.repository.DemandAnd
 import org.eclipse.tractusx.puris.backend.common.edc.domain.repository.DemandContractMappingRepository;
 import org.eclipse.tractusx.puris.backend.common.edc.domain.repository.DtrContractMappingRepository;
 import org.eclipse.tractusx.puris.backend.common.edc.domain.repository.GeneralContractMappingRepository;
+import org.eclipse.tractusx.puris.backend.common.edc.domain.repository.ItemStockAnonymizedContractMappingRepository;
 import org.eclipse.tractusx.puris.backend.common.edc.domain.repository.ItemStockContractMappingRepository;
 import org.eclipse.tractusx.puris.backend.common.edc.domain.repository.PartTypeContractMappingRepository;
 import org.eclipse.tractusx.puris.backend.common.edc.domain.repository.ProductionContractMappingRepository;
@@ -38,7 +41,7 @@ import org.eclipse.tractusx.puris.backend.masterdata.domain.model.Partner;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.lang.reflect.Constructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @Slf4j
@@ -65,6 +68,9 @@ public class EdcContractMappingService {
     @Autowired
     private DaysOfSupplyContractMappingRepository daysOfSupplyContractMappingRepository;
 
+    @Autowired
+    private ItemStockAnonymizedContractMappingRepository itemStockAnonymizedContractMappingRepository;
+  
     @Autowired
     private DeliveryAnonymizedContractMappingRepository deliveryAnonymizedContractMappingRepository;
 
@@ -133,6 +139,7 @@ public class EdcContractMappingService {
             case DELIVERY_SUBMODEL -> deliveryContractMappingRepository;
             case NOTIFICATION -> demandAndCapacityNotificationContractMappingRepository;
             case DAYS_OF_SUPPLY -> daysOfSupplyContractMappingRepository;
+            case ITEM_STOCK_ANONYMIZED_SUBMODEL -> itemStockAnonymizedContractMappingRepository;
             case DELIVERY_ANONYMIZED_SUBMODEL -> deliveryAnonymizedContractMappingRepository;
             case PART_TYPE_INFORMATION_SUBMODEL -> partTypeContractMappingRepository;
         };